+++ julesTk/view/__init__.py
@@ -1,16 +1,24 @@
 """Implementing basic view objects"""
 
-from julesTk import *
+import sys
+if sys.version_info[0] < 3:
+    import Tkinter as tk
+    import ttk
+else:
+    import tkinter as tk
+    from tkinter import ttk
 
 __author__ = "Joeri Jongbloets <joeri@jongbloets.net>"
 
 
-class Frame(ttk.Frame, object):
+class BaseFrame(object):
+    """Empty base class used to share most basic functions between frames and views"""
 
     FONT_FAMILY = "Verdana"
     FONT_SMALL = (FONT_FAMILY, 8)
     FONT_NORMAL = (FONT_FAMILY, 10)
     FONT_LARGE = (FONT_FAMILY, 14)
+    FONT_HUGE = (FONT_FAMILY, 16)
 
     @staticmethod
     def configure_grid(w, sticky="nsew", **kwargs):
@@ -32,30 +40,48 @@ class Frame(ttk.Frame, object):
         for index in indexes:
             w.grid_rowconfigure(index, **kwargs)
 
+
+class Frame(ttk.Frame, BaseFrame):
+
     def __init__(self, parent):
-        """
-
-        :rtype: Tkinter.Frame
-        """
-        ttk.Frame.__init__(self, parent)
-        self._parent = parent
-
-    @property
-    def parent(self):
-        return self._parent
-
-    def setup(self):
-        raise NotImplementedError
+        super(Frame, self).__init__(parent)
 
 
-class View(Frame):
-    """Interacting frame"""
+class BaseView(BaseFrame):
+    """Empty Base class used for checking if an object belongs to the view family"""
 
     def __init__(self, parent, controller):
-        super(View, self).__init__(parent)
+        """Initialize a BaseView
+
+        """
+        super(BaseView, self).__init__()
+        self._configured = False
+        self._parent = parent
         self._controller = controller
         self._variables = {}
         self._widgets = {}
+
+    def __del__(self):
+        self.close()
+
+    @property
+    def parent(self):
+        """Return the master of this widget
+
+        :rtype: julesTk.view.BaseView
+        """
+        return self._parent
+
+    @property
+    def application(self):
+        """Return the root level widget
+
+        :rtype: Tkinter.Tk or tkinter.Tk
+        """
+        result = self.parent
+        if not isinstance(result, tk.Tk):
+            result = self.parent.parent
+        return result
 
     @property
     def controller(self):
@@ -67,203 +93,169 @@ class View(Frame):
 
     @property
     def variables(self):
-        """Dictionary of variables registered to this view
+        """Variables registered to this view
 
-        :rtype: dict[str, tk.Variable]
+        :rtype: dict[str, Tkinter.Variable | tkinter.Variable]
         """
         return self._variables
 
-    def has_variable(self, name):
-        """Returns whether a variable is registered to this view using the given name
+    @property
+    def widgets(self):
+        """Widgets registered to this view
 
+        :rtype: dict[str, Tkinter.BaseWidget | tkinter.BaseWidget]
+        """
+        return self._widgets
+
+    def prepare(self):
+        self._prepare()
+        self._configured = True
+        return self
+
+    def _prepare(self):
+        raise NotImplementedError
+
+    def show(self):
+        if not self._configured:
+            self.prepare()
+        self._show()
+        return self
+
+    def _show(self):
+        raise NotImplementedError
+
+    def hide(self):
+        self._hide()
+
+    def _hide(self):
+        raise NotImplementedError
+
+    def close(self):
+        self._close()
+
+    def _close(self):
+        raise NotImplementedError
+
+    def has_variable(self, name):
+        """Whether a variable is registered under the given name
+
+        :param name: Name to search for
+        :type name: str
         :rtype: bool
         """
         return name in self.variables.keys()
 
-    def get_variable(self, name):
-        """Returns a variable registered to this view
-
-        :param name: Name of the variable
-        :type name: str
-        :return: The variable
-        :rtype: tk.Variable
-        """
-        if not self.has_variable(name):
-            raise KeyError("Unknown variable: {}".format(name))
-        return self.variables[name]
-
     def add_variable(self, name, variable):
-        """Registers a variable under a new name to this view
+        """Register a variable under a new name to the view
 
-        :param name: The name under which the variable should be registered
+        :param name: Name to register the variable under
         :type name: str
         :param variable: The variable to register
-        :type variable: tk.Variable
-        :return: The registered variable
-        :rtype: tk.Variable
+        :type variable: Tkinter.Variable | tkinter.Variable
+        :rtype: Tkinter.Variable | tkinter.Variable
         """
         if self.has_variable(name):
-            raise KeyError("Already registered a variable as: {}".format(name))
+            raise KeyError("Variable name already registered to this view")
         self.variables[name] = variable
         return variable
 
-    def remove_variable(self, name):
-        """Removes a variable (name) from the registry
+    def get_variable(self, name):
+        """Get the variable registered to this view under the given name
 
-        :return: Whether the variable is still in the registry
-        :rtype: bool
+        :param name: Name of the variable
+        :type name: str
+        :return: The requested variable
+        :rtype: Tkinter.Variable | tkinter.Variable
         """
         if not self.has_variable(name):
-            raise KeyError("Unknown variable: {}".format(name))
+            raise KeyError("Variable name not registered to this view")
+        return self.variables[name]
+
+    def remove_variable(self, name):
+        """Remove a variable from the registry of this view
+
+        The variable is NOT destroyed!
+
+        :param name: Name of the variable to remove
+        :type name: str
+        :return: The removed variable
+        :rtype: Tkinter.Variable | tkinter.Variable
+        """
+        if not self.has_variable(name):
+            raise KeyError("Variable name not registered to this view")
         self.variables.pop(name)
         return not self.has_variable(name)
 
-    @property
-    def widgets(self):
-        """Dictionary of widgets registered to this view
-
-        :rtype: dict[str, tk.Widget]
-        """
-        return self._widgets
-
     def has_widget(self, name):
-        """Whether this view has a widget registered under the given name
+        """Whether a widget is registered using the given name
 
-        :param name: Name to look for
         :type name: str
         :rtype: bool
         """
         return name in self.widgets.keys()
 
-    def get_widget(self, name):
-        """Returns the widget registered to this view under the given name
-
-        :param name: Name to look for
-        :type name: str
-        :rtype: tk.Widget
-        """
-        if not self.has_widget(name):
-            raise KeyError("Unknown widget: {}".format(name))
-        return self.widgets[name]
-
     def add_widget(self, name, widget):
-        """Registers a widget under a new name to the view
+        """Register a new widget to this view
 
-        :raises: KeyError, if name is already used
-
-        :param name: Name to register the widget under
+        :param name: Name of the widget
         :type name: str
         :param widget: The widget to register
-        :type widget: tk.Widget
-        :return: The widget that as registered
-        :rtype: tk.Widget
+        :type widget: Tkinter.BaseWidget | tkinter.BaseWidget
+        :return: The registered widget
+        :rtype: Tkinter.BaseWidget | tkinter.BaseWidget
         """
         if self.has_widget(name):
-            raise KeyError("Already registered a widget as: {}".format(name))
+            raise KeyError("Widget name already registered to this view")
         self.widgets[name] = widget
         return widget
 
-    def remove_widget(self, name):
-        """Removes a widget (name) from the registry
+    def get_widget(self, name):
+        """Get the widget registered under the given name
 
-        :param name: Name to remove
+        :param name: Name of the widget
         :type name: str
-        :return: Whether the widget is still registered
-        :rtype: bool
+        :return: The requested widget
+        :rtype: Tkinter.BaseWidget | tkinter.BaseWidget
         """
         if not self.has_widget(name):
-            raise KeyError("Unknown widget: {}".format(name))
-        self.widgets.pop(name)
-        return name not in self.widgets.keys()
+            raise KeyError("Widget name not registered to this view")
+        return self.widgets[name]
 
-    def show(self):
+    def remove_widget(self, name):
+        """Remove a widget from the registry of this view
+
+        The widget is NOT destroyed
+
+        :param name: Name of the widget to remove
+        :type name: str
+        :return: The widget that was removed
+        :rtype: Tkinter.BaseWidget | tkinter.BaseWidget
+        """
+        if not self.has_widget(name):
+            raise KeyError("Widget name not registered to this view")
+        self.widgets.pop(name)
+        return not self.has_widget(name)
+
+
+class View(Frame, BaseView):
+    """A Frame acting as """
+
+    def __init__(self, parent, controller):
+        Frame.__init__(self, parent)
+        BaseView.__init__(self, parent, controller)
+
+    def _show(self):
         """Shows the view"""
         self.tkraise()
 
-    def hide(self):
+    def _hide(self):
         """Hides the view (temporarily)"""
         self.grid_remove()
 
-    def close(self):
+    def _close(self):
         """Closes the view"""
-        return True
+        self.destroy()
 
-    def setup(self):
+    def _prepare(self):
         """Configure this view"""
         raise NotImplementedError
-
-
-class ViewSet(View):
-    """A view can contain one or more views"""
-
-    def __init__(self, parent, controller):
-        super(ViewSet, self).__init__(parent, controller)
-        self._views = {}
-
-    @property
-    def views(self):
-        """
-
-        :rtype: dict[str, pyLabJackView.view.View]
-        """
-        return self._views
-
-    def get_view(self, name):
-        """
-
-        :param name: Name of the view
-        :rtype: julesTk.view.View
-        """
-        if not self.has_view(name):
-            KeyError("No view registered under: {}".format(name))
-        return self.views[name]
-
-    def has_view(self, name):
-        return name in self.views
-
-    def add_view(self, name, frame):
-        if self.has_view(name):
-            KeyError("Already registered a view under: {}".format(name))
-        self.views[name] = frame
-
-    def remove_view(self, name):
-        if not self.has_view(name):
-            KeyError("No view registered under: {}".format(name))
-        self.get_view(name).close()
-        self.views.pop(name)
-
-    def show_view(self, name):
-        """Show a specific View"""
-        v = self.get_view(name)
-        v.show()
-
-    def hide_views(self):
-        """Hide all Views"""
-        for v in self._views.keys():
-            self.hide_view(v)
-        # done
-
-    def hide_view(self, name):
-        """Hide a specific View"""
-        v = self.get_view(name)
-        v.grid_remove()
-
-    def show(self):
-        """Show this ViewSet"""
-        raise NotImplementedError
-
-    def hide(self):
-        """Hide this ViewSet"""
-        raise NotImplementedError
-
-    def setup(self):
-        """Configure this ViewSet"""
-        raise NotImplementedError
-
-    def close(self):
-        """Close this ViewSet"""
-        while len(self.views.values()) > 0:
-            v = self.views.keys()[0]
-            self.get_view(v).close()
-            self.remove_view(v)
-        super(ViewSet, self).close()